Refactor: 기존 사용하던 StatusCode.RESPONSE_SUCCESS에서 새로 생성한 GeneralCode.RESPONSE_SUCCESS로 변경
