+++ be/src/main/java/codesquad/secondhand/controller/CategoryController.java
@@ -1,6 +1,5 @@
 package codesquad.secondhand.controller;
 
-import codesquad.secondhand.dto.ResponseDto;
 import codesquad.secondhand.dto.ResponseListDto;
 import codesquad.secondhand.dto.category.CategoryDto;
 import codesquad.secondhand.service.CategoryService;
@@ -13,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-import static codesquad.secondhand.dto.StatusCode.RESPONSE_SUCCESS;
+import static codesquad.secondhand.exception.code.GeneralCode.RESPONSE_SUCCESS;
 
 @Slf4j
 @RestController
