+++ be/src/main/java/codesquad/secondhand/controller/LocationController.java
@@ -12,8 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-import static codesquad.secondhand.dto.StatusCode.RESPONSE_SUCCESS;
-
+import static codesquad.secondhand.exception.code.GeneralCode.RESPONSE_SUCCESS;
 @Slf4j
 @RestController
 @RequiredArgsConstructor
