+++ be/src/main/java/codesquad/secondhand/controller/MemberController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
 
-import static codesquad.secondhand.dto.StatusCode.RESPONSE_SUCCESS;
+import static codesquad.secondhand.exception.code.GeneralCode.RESPONSE_SUCCESS;
 
 @Slf4j
 @RestController
